changed prepareTokensData file to ts
